loading screen on scanner
