+++ scanner/src/routes/main/+layout.ts
@@ -1,0 +1,1 @@
+export const prerender = true
