+++ scanner/src/routes/main/+page.server.ts
