+++ scanner/src/routes/main/scan/+page.svelte
