+++ scanner/src/routes/style.css
@@ -15,3 +15,11 @@ div:has(> video) {
 video {
   object-fit: cover;
 }
+.background {
+  background: rgba(16, 25, 32, 1);
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  top: 0;
+  left: 0;
+}
